Feat/kube chart deploy

+++ back-end/src/routers/iam.py
@@ -64,7 +64,9 @@ async def get_current_user(
         token_data = TokenData(userid=userid)
     except JWTError:
         raise credentials_exception
-    user = await db["users"].find_one({"userid": token_data.userid})
+    async with await mongo_client.start_session() as session:
+        async with session.start_transaction():
+            user = await db["users"].find_one({"userid": token_data.userid})
     if user is None:
         raise credentials_exception
     return user
@@ -79,17 +81,10 @@ async def add_user(
     db, mongo_client = db
     try:
         item.password = get_password_hash(item.password)
-        user = await db["users"].insert_one(
-            {
-                "userid": item.userid,
-                "name": item.name,
-                "password": item.password,
-                "admin_priv": item.admin_priv,
-            }
-        )
-        add_user = await db["users"].find_one(
-            {"_id": user.inserted_id}, {"_id": False, "password": False}
-        )
+        async with await mongo_client.start_session() as session:
+            async with session.start_transaction():
+                user = await db["users"].insert_one({"userid": item.userid,"name": item.name,"password": item.password,"admin_priv":item.admin_priv})
+                add_user = await db["users"].find_one({"_id": user.inserted_id},{"_id":False,"password":False})
         print(add_user)
         return JSONResponse(
             status_code=status.HTTP_201_CREATED,
@@ -118,9 +113,9 @@ async def delete_user(
 ):
     db, mongo_client = db
     try:
-        delete_result = await db["users"].delete_many(
-            {"userid": {"$in": userid}}
-        )
+        async with await mongo_client.start_session() as session:
+            async with session.start_transaction():
+                delete_result = await db["users"].delete_many({"userid": {"$in": userid}})
         return Response(status_code=204)
     except:
         raise HTTPException(status_code=404, detail=f"Not found")
@@ -135,17 +130,12 @@ async def update_user(
     db, mongo_client = db
     try:
         user.password = get_password_hash(user.password)
-        update_result = await db["users"].update_one(
-            {"userid": user.userid},
-            {
-                "$set": {
-                    "userid": user.userid,
-                    "name": user.name,
-                    "password": user.password,
-                    "admin_priv": user.admin_priv,
-                }
-            },
-        )
+        async with await mongo_client.start_session() as session:
+            async with session.start_transaction():
+                update_result = await db["users"].update_one(
+                    {"userid": user.userid},
+                    {"$set": {"userid": user.userid,"name": user.name,"password": user.password,"admin_priv":user.admin_priv}},
+                    )
     except ValueError as e:
         raise HTTPException(status_code=422, detail=f"")
     except:
@@ -158,37 +148,34 @@ async def auth_user_admin(
     form_data: OAuth2PasswordRequestForm = Depends(), db=Depends(get_db)
 ):
     db, mongo_client = db
-    if (
-        user := await db["users"].find_one({"userid": form_data.username})
-    ) is not None:
-        if verify_password(form_data.password, user["password"]) is True:
-            if user["admin_priv"] is True:
-                access_token_expires = timedelta(
-                    minutes=ACCESS_TOKEN_EXPIRE_MINUTES
-                )
-                access_token = create_access_token(
-                    data={"sub": user["userid"]},
-                    expires_delta=access_token_expires,
-                )
-                return {"access_token": access_token, "token_type": "bearer"}
+    async with await mongo_client.start_session() as session:
+        async with session.start_transaction():
+            if (user := await db["users"].find_one({"userid": form_data.username})) is not None:
+                if verify_password(form_data.password, user["password"]) is True:
+                    if user["admin_priv"] is True:
+                        access_token_expires = timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
+                        access_token = create_access_token(
+                            data={"sub": user["userid"]}, expires_delta=access_token_expires
+                        )
+                        return {"access_token": access_token, "token_type": "bearer"}
+                    else:
+                        raise HTTPException(
+                            status_code=401,
+                            detail=f"User is not admin",
+                            headers={"WWW-Authenticate": "Bearer"},
+                        )
+                else:
+                    raise HTTPException(
+                        status_code=401,
+                        detail=f"Password is wrong",
+                        headers={"WWW-Authenticate": "Bearer"},
+                    )
             else:
                 raise HTTPException(
-                    status_code=401,
-                    detail=f"User is not admin",
+                    status_code=404,
+                    detail=f"User ID does not exist",
                     headers={"WWW-Authenticate": "Bearer"},
                 )
-        else:
-            raise HTTPException(
-                status_code=401,
-                detail=f"Password is wrong",
-                headers={"WWW-Authenticate": "Bearer"},
-            )
-    else:
-        raise HTTPException(
-            status_code=404,
-            detail=f"User ID does not exist",
-            headers={"WWW-Authenticate": "Bearer"},
-        )
 
 
 # check whether user is authenticated and has admin priv
@@ -205,53 +192,33 @@ async def check_user_admin(current_user: User = Depends(get_current_user)):
 
 # list users with pagination
 @router.post("/{page_num}")
-async def get_users(
-    pages_user: UserPage,
-    page_num: int = Path(ge=1),
-    current_user: User = Depends(get_current_user),
-    db=Depends(get_db),
-):
-    db, mongo_client = db
-    if current_user["admin_priv"] is True:
-        try:
-            # check number of documents to skip past
-            skips = pages_user.user_num * (page_num - 1)
-            # lookups for name and admin priv matching
-            lookup = {}
-            if pages_user.name != None:
-                lookup["name"] = {"$regex": pages_user.name, "$options": "i"}
-            if pages_user.admin_priv != None:
-                lookup["admin_priv"] = pages_user.admin_priv
-            # dont skip if 1st page
-            if skips <= 0:
-                # find from users in MongodDB exclude ObjectID
-                cursor = (
-                    db["users"]
-                    .find(lookup, {"_id": False, "password": False})
-                    .limit(pages_user.user_num)
-                )
-            # else call cursor with skips
-            else:
-                # find from users in MongodDB exclude ObjectID
-                cursor = (
-                    db["users"]
-                    .find(lookup, {"_id": False, "password": False})
-                    .skip(skips)
-                    .limit(pages_user.user_num)
-                )
-            # get from MongoDB and convert to list
-            cursor = await cursor.to_list(length=pages_user.user_num)
-            # return documents
-            return JSONResponse(status_code=status.HTTP_200_OK, content=cursor)
-        except ValueError as e:
-            return JSONResponse(
-                status_code=status.HTTP_422_UNPROCESSABLE_ENTITY, content=f""
-            )
-        except:
-            return HTTPException(status_code=404)
-    # ensure when loading this endpoint that user is authenticated as admin",headers={"WWW-Authenticate": "Bearer"})
-    raise HTTPException(
-        status_code=401,
-        detail=f"User is not admin",
-        headers={"WWW-Authenticate": "Bearer"},
-    )
+async def get_users(pages_user : UserPage, page_num: int = Path(ge = 1),current_user: User = Depends(get_current_user), db=Depends(get_db) ):
+        db, mongo_client = db
+        if current_user["admin_priv"] is True:
+            try:
+                # check number of documents to skip past
+                skips = pages_user.user_num * (page_num - 1)
+                # lookups for name and admin priv matching
+                lookup = {}
+                if pages_user.name != None:
+                    lookup['name'] = { '$regex' : pages_user.name, '$options' : 'i' }
+                if pages_user.admin_priv != None:
+                    lookup['admin_priv'] = pages_user.admin_priv
+                # dont skip if 1st page
+                async with await mongo_client.start_session() as session:
+                    async with session.start_transaction():
+                        if skips <= 0: 
+                            # find from users in MongodDB exclude ObjectID and convert to list
+                            cursor = await (db['users'].find(lookup, {'_id': False, 'password' : False}).limit(pages_user.user_num)).to_list(length = pages_user.user_num)
+                        # else call cursor with skips
+                        else:
+                            # find from users in MongodDB exclude ObjectID and convert to list
+                            cursor =  await (db['users'].find(lookup, {'_id': False, 'password' : False}).skip(skips).limit(pages_user.user_num)).to_list(length = pages_user.user_num)
+                        # return documents
+                return JSONResponse(status_code=status.HTTP_200_OK, content=cursor)
+            except ValueError as e:
+                return JSONResponse(status_code=status.HTTP_422_UNPROCESSABLE_ENTITY, content=f"")
+            except:
+                return HTTPException(status_code=404)
+        # ensure when loading this endpoint that user is authenticateds not admin",headers={"WWW-Authenticate": "Bearer"})
+        raise HTTPException(status_code = 401,detail=f"User is not admin",headers={"WWW-Authenticate": "Bearer"})
